khonsu::distributed::paxos_service::paxos_service_server

Trait PaxosService

pub trait PaxosService:
    Send
    + Sync
    + 'static {
    // Required method
    fn send_message<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ConsensusMessage>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<SendMessageResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Available on crate feature distributed only.

Generated trait containing gRPC methods that should be implemented for use with PaxosServiceServer.

Required Methods

fn send_message<'life0, 'async_trait>( &'life0 self, request: Request<ConsensusMessage>, ) -> Pin<Box<dyn Future<Output = Result<Response<SendMessageResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

A simple RPC for sending a message to another Multi Paxos node.