khonsu::distributed

Struct ReplicatedCommit

pub struct ReplicatedCommit {
    pub transaction_id: GlobalTransactionId,
    pub write_set: HashMap<String, SerializableVersionedValue>,
    pub state: TransactionState,
    pub read_set: HashMap<String, u64>,
    pub prepare_timestamp: u64,
    pub decision_timestamp: Option<u64>,
    pub coordinator_node: NodeId,
    pub participant_nodes: Vec<NodeId>,
}

Available on crate feature distributed only.

Represents a transaction commit that is replicated through OmniPaxos.

Fields
transaction_id: GlobalTransactionId

The globally unique identifier for this transaction.

write_set: HashMap<String, SerializableVersionedValue>

The set of changes to be applied by this transaction.

state: TransactionState

The current state of the transaction in the 2PC protocol.

read_set: HashMap<String, u64>

The read set of the transaction, used for validation during recovery.

prepare_timestamp: u64

The timestamp when this transaction was prepared.

decision_timestamp: Option<u64>

The timestamp when this transaction was committed or aborted.

coordinator_node: NodeId

The node that coordinated this transaction.

participant_nodes: Vec<NodeId>

The set of nodes that participated in this transaction.

Implementations

impl ReplicatedCommit

pub fn new_prepared( transaction_id: GlobalTransactionId, write_set: HashMap<String, SerializableVersionedValue>, read_set: HashMap<String, u64>, prepare_timestamp: u64, coordinator_node: NodeId, participant_nodes: Vec<NodeId>, ) -> Self

Creates a new ReplicatedCommit in the Prepared state.

pub fn mark_committed(&mut self, decision_timestamp: u64)

Marks this transaction as committed.

pub fn mark_aborted(&mut self, decision_timestamp: u64)

Marks this transaction as aborted.

pub fn is_committed(&self) -> bool

Returns whether this transaction is committed.

pub fn is_aborted(&self) -> bool

Returns whether this transaction is aborted.

pub fn is_prepared(&self) -> bool

Returns whether this transaction is prepared but not yet decided.