khonsu

Module prelude

Prelude of the Khonsu.

Re-exports

pub use crate::khonsu::Khonsu;

pub use crate::transaction::Transaction;

pub use crate::conflict::detection::*;

pub use crate::conflict::resolution::*;

pub use crate::data_store::txn_buffer::*;

pub use crate::data_store::versioned_value::*;

pub use crate::dependency_tracking::*;

pub use crate::errors::*;

pub use crate::isolation::*;

pub use crate::storage::*;

pub use crate::distributed::*;distributed

Structs

ClusterConfigdistributed

Configuration for an OmniPaxos cluster.

FlexibleQuorumdistributed

Flexible quorums can be used to increase/decrease the read and write quorum sizes, for different latency vs fault tolerance tradeoffs.

LogSyncdistributed

Struct used to help another server synchronize their log with the current state of our own log.

SequenceNumberdistributed

Keeps track of the ordering of messages in the accept phase

SnapshottedEntrydistributed

Enums

LogEntrydistributed

The entry read in the log.

Constants

READ_ERROR_MSGdistributed

Error message to display when there was an error reading to the storage implementation.

WRITE_ERROR_MSGdistributed

Error message to display when there was an error writing to the storage implementation.

Type Aliases

ConfigurationIddistributed

ID for an OmniPaxos configuration (i.e., the set of servers in an OmniPaxos cluster)

NodeIddistributed

ID for an OmniPaxos node

TrimmedIndexdistributed

Derive Macros

Entrydistributed

Derive macro for declaring an OmniPaxos log entry type.

UniCacheEntrydistributed

Derive macro for declaring an OmniPaxos log entry type that uses UniCache. UniCache can reduce the amount of transmitted data by caching the values of user-specified fields and encoding them as smaller types. For instance, a popular String that appears repeatedly can be sent over the network as an u8.