khonsu::prelude

Derive Macro UniCacheEntry

#[derive(UniCacheEntry)]
{
    // Attributes available to this derive:
    #[unicache]
    #[snapshot]
}

Available on crate feature distributed only.

Derive macro for declaring an OmniPaxos log entry type that uses UniCache. UniCache can reduce the amount of transmitted data by caching the values of user-specified fields and encoding them as smaller types. For instance, a popular String that appears repeatedly can be sent over the network as an u8.

Attributes

encoding(T): (Optional) The type for what the annotated field should be encoded as. The default is u8.

size(usize): (Optional) The size of the cache for this field. Should not be larger than the max size of the encoding type, e.g., if encoding(u8) is used, the max size should be 255. Default is u8::MAX.

cache(C): (Optional) The cache implementation which is a type C: UniCache. cache(lru) or cache(lfu) are provided out of the box. The default is lru (least-recently-used) eviction policy.

Usage

use omnipaxos_macros::UniCacheEntry;
#[derive(UniCacheEntry, Clone, Debug, Default, PartialOrd, PartialEq, Serialize, Deserialize, Eq, Hash)]
pub struct Person {
    pub id: u64,
    #[unicache(encoding(u8))]   // First names can be cached e.g., if "John" repeatedly occurs in the log, it will be sent as a u8 instead.
    first_name: String,
    #[unicache(encoding(u64), size(100))]   // The size of the cache for this field can be optionally specified using the `size` attribute.
    last_name: String,
    #[unicache(encoding(u64), size(20), cache(lru))]    // use provided UniCache implementation with least-recently-used (LRU) eviction policy.
    job: String,
}